pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^5.0"
python = "^3.10"
python-decouple = "^3.8"
aiohttp = "^3.12.15"
requests = "^2.32.5"
cassandra-driver = "^3.29.2"
tenacity = "^9.1.2"
sentry-sdk = "^2.38.0"
django-cassandra-engine = "^1.10.0"
pandas = "^2.3.2"
stellar-sdk = "^13.0.0"
gunicorn = "^23.0.0"
whitenoise = "^6.11.0"
google-cloud-bigquery = "^3.38.0"
django-ratelimit = "^4.1.0"
redis = "^6.4.0"
apscheduler = "^3.11.0"
[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"
pytest-xdist = "^3.3.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "StellarMapWeb.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*", "*Tests", "*TestCase"]
python_functions = ["test_*"]
testpaths = ["apiApp/tests", "webApp/tests"]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (database, API, external services)",
    "e2e: End-to-end tests (full workflow tests)",
    "slow: Tests that take significant time",
    "performance: Performance and optimization tests",
    "regression: Regression tests for bug fixes"
]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--tb=short",
    "-p no:warnings",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[tool.coverage.run]
source = ["apiApp", "webApp"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.pythonlibs/*"
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"